molecule_platform-3/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
rdkit==2023.9.5
numpy==1.26.4
pandas==2.2.0
requests==2.31.0
python-multipart==0.0.9
pydantic==2.6.1
httpx==0.26.0
PyPDF2==3.0.1
chembl_webresource_client==0.10.8
pubchempy==1.0.4
biopython==1.83
IPython==8.21.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
aiohttp==3.9.3
python-dotenv==1.0.0
